[#2] move directory
